Rename modal directory and set up subfolders
