+++ components/admin/projects/ProjectEdit/EditVideoModal/ModalSections/FileSidebar/FileSidebar.scss
